# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Workforce scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workplaces:
    get:
      operationId: listWorkplaces
      tags: [workplaces]
      summary: List workplaces
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workplace"
    post:
      operationId: createWorkplace
      tags: [workplaces]
      summary: Create a workplace (admin/manager)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkplaceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workplace"
  /workplaces/{id}:
    get:
      operationId: getWorkplace
      tags: [workplaces]
      summary: Get a workplace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workplace"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateWorkplace
      tags: [workplaces]
      summary: Update a workplace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkplaceRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workplace"
    delete:
      operationId: deleteWorkplace
      tags: [workplaces]
      summary: Delete a workplace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted

  # ── Shifts ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ────────────────────────────────────────────────────────────────
...
    Workplace:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
        radiusMeters:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, name, radiusMeters]
    WorkplaceRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
        radiusMeters:
          type: integer
      required: [name]

    # ── Shift ────────────────────────────────────────────────────────────────
...